Basic Stack Implementation
This creates the concept of a "stack" that groups and renders multiple
maps together.  This implementation is a little bit clunky because it
special-cases them in rendering and focus operations.  I may refactor
the base component to nicely handle multiple types of renderable
objects at some point, but I'm guessing I'l probably just deal with
this clunkiness for now until I've got a working solution.